requirements.txt
mesa==3.3.1
opendssdirect.py==0.9.4
pulp==3.3.0
mpi4py==4.1.1
plotly==5.24.1
dash==2.18.1
dash-bootstrap-components==1.6.0
pandas==2.2.2
numpy==1.26.4
matplotlib==3.9.0
scipy